renamed decorators.py back to registry.py
