+++ goodtables/checks/extra_header.py
@@ -7,7 +7,7 @@ from __future__ import unicode_literals
 from copy import copy
 from jsontableschema import Schema, infer
 from ..spec import spec
-from ..decorators import check
+from ..registry import check
 
 
 # Module API
