+++ goodtables/checks/missing_value.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 
 from copy import copy
 from ..spec import spec
-from ..decorators import check
+from ..registry import check
 
 
 # Module API
