+++ goodtables/presets/nested.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from copy import deepcopy
-from ..decorators import preset
+from ..registry import preset
 from .. import exceptions
