+++ goodtables/registry.py
